recognize streaming rpc pattern
The captures `rpc.request.protobuf` and `rpc.response.protobuf`
captures can now be prefixed with “stream “.

This adds support for streaming rpc calls as defined by the gRPC
project (which is developed in conjunction with protocol buffers). This
behavior is modeled in gRPC’s docs:
http://www.grpc.io/docs/guides/concepts.html

A test case is also added to `test.proto`.